Update 671. Second Minimum Node In a Binary Tree .cpp

+++ Binary Tree/671. Second Minimum Node In a Binary Tree .cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     
-    void find(TreeNode *root, int &min_val , int &ans)
+    void find(TreeNode *root, int &min_val , long &ans)
     {
         if(!root)
             return ;
@@ -20,10 +20,10 @@ public:
     int findSecondMinimumValue(TreeNode* root) 
     {
         int min_val = root->val;
-        int ans = INT_MAX;
+        long ans = LONG_MAX;
         
         find(root, min_val, ans);
         
-        return ans < INT_MAX ? ans : -1;
+        return ans < LONG_MAX ? ans : -1;
     }
 };
